+++ ragas/tests/unit/backends/test_local_csv.py
@@ -1,14 +1,13 @@
-"""Comprehensive tests for LocalJSONLBackend to test serialization capabilities."""
+"""Comprehensive tests for LocalCSVBackend to test serialization edge cases."""
 
 import tempfile
 from datetime import datetime, date
 from pathlib import Path
-import typing as t
 from typing import List, Dict, Any, Optional
 import pytest
-from pydantic import BaseModel
+from pydantic import BaseModel, ValidationError
 
-from ragas.experimental.backends.local_jsonl import LocalJSONLBackend
+from ragas.backends.local_csv import LocalCSVBackend
 
 
 # Test BaseModel classes
@@ -41,10 +40,10 @@ def temp_dir():
         yield tmp_dir
 
 
-@pytest.fixture(name="backend")
-def jsonl_backend_fixture(temp_dir):
-    """Create a LocalJSONLBackend instance with temp directory."""
-    return LocalJSONLBackend(temp_dir)
+@pytest.fixture
+def backend(temp_dir):
+    """Create a LocalCSVBackend instance with temp directory."""
+    return LocalCSVBackend(temp_dir)
 
 
 @pytest.fixture
@@ -99,11 +98,11 @@ def nested_data():
 
 # 1. Basic Functionality Tests
 class TestBasicFunctionality:
-    """Test basic LocalJSONLBackend functionality."""
+    """Test basic LocalCSVBackend functionality."""
 
     def test_initialization(self, temp_dir):
         """Test backend initialization."""
-        backend = LocalJSONLBackend(temp_dir)
+        backend = LocalCSVBackend(temp_dir)
         assert backend.root_dir == Path(temp_dir)
 
     def test_get_data_dir(self, backend):
@@ -119,8 +118,8 @@ class TestBasicFunctionality:
         dataset_path = backend._get_file_path("datasets", "test_dataset")
         experiment_path = backend._get_file_path("experiments", "test_experiment")
 
-        assert dataset_path.name == "test_dataset.jsonl"
-        assert experiment_path.name == "test_experiment.jsonl"
+        assert dataset_path.name == "test_dataset.csv"
+        assert experiment_path.name == "test_experiment.csv"
 
     def test_save_and_load_simple_data(self, backend, simple_data):
         """Test basic save and load cycle with simple data."""
@@ -130,12 +129,11 @@ class TestBasicFunctionality:
         # Load dataset
         loaded_data = backend.load_dataset("test_simple")
 
-        # Verify data structure - JSONL should preserve types
+        # Verify data structure (note: all values become strings in CSV)
         assert len(loaded_data) == len(simple_data)
         assert loaded_data[0]["name"] == "Alice"
-        assert loaded_data[0]["age"] == 30  # Should be int, not string
-        assert loaded_data[0]["score"] == 85.5  # Should be float, not string
-        assert loaded_data[0]["is_active"] is True  # Should be bool, not string
+        # This will fail because CSV converts everything to strings
+        # assert loaded_data[0]["age"] == 30  # This will be "30"
 
     def test_directory_creation(self, backend, simple_data):
         """Test automatic directory creation."""
@@ -186,10 +184,11 @@ class TestBasicFunctionality:
 
 # 2. Data Type Edge Cases (The Real Challenge)
 class TestDataTypeEdgeCases:
-    """Test complex data types that JSONL should handle properly."""
+    """Test complex data types that reveal CSV serialization issues."""
 
+    @pytest.mark.skip(reason="CSV backend doesn't support nested dictionaries")
     def test_nested_dictionaries(self, backend):
-        """Test nested dictionary serialization - JSONL should handle this."""
+        """Test nested dictionary serialization - THIS SHOULD FAIL."""
         data = [
             {
                 "id": 1,
@@ -201,15 +200,20 @@ class TestDataTypeEdgeCases:
         backend.save_dataset("nested_test", data)
         loaded_data = backend.load_dataset("nested_test")
 
-        # JSONL should preserve nested dictionaries exactly
+        # This will fail - nested dicts become string representations
         assert loaded_data[0]["metadata"] == {
             "score": 0.85,
             "tags": ["test", "important"],
         }
-        assert loaded_data[0]["config"]["settings"]["temperature"] == 0.7
 
+        # Show what actually happens
+        print(f"Original: {data[0]['metadata']}")
+        print(f"Loaded: {loaded_data[0]['metadata']}")
+        print(f"Type: {type(loaded_data[0]['metadata'])}")
+
+    @pytest.mark.skip(reason="CSV backend doesn't support lists of objects")
     def test_lists_of_objects(self, backend):
-        """Test lists of objects serialization - JSONL should handle this."""
+        """Test lists of objects serialization - THIS SHOULD FAIL."""
         data = [
             {
                 "id": 1,
@@ -223,14 +227,17 @@ class TestDataTypeEdgeCases:
         backend.save_dataset("list_test", data)
         loaded_data = backend.load_dataset("list_test")
 
-        # JSONL should preserve lists of objects
+        # This will fail - lists become string representations
         assert loaded_data[0]["results"][0]["metric"] == "accuracy"
-        assert loaded_data[0]["results"][0]["value"] == 0.9
-        assert loaded_data[0]["results"][1]["metric"] == "precision"
-        assert loaded_data[0]["results"][1]["value"] == 0.8
 
+        # Show what actually happens
+        print(f"Original: {data[0]['results']}")
+        print(f"Loaded: {loaded_data[0]['results']}")
+        print(f"Type: {type(loaded_data[0]['results'])}")
+
+    @pytest.mark.skip(reason="CSV backend doesn't preserve data types")
     def test_mixed_types(self, backend):
-        """Test mixed data types - JSONL should preserve all types."""
+        """Test mixed data types - THIS WILL PARTIALLY FAIL."""
         data = [
             {
                 "str_field": "text",
@@ -244,15 +251,19 @@ class TestDataTypeEdgeCases:
         backend.save_dataset("mixed_test", data)
         loaded_data = backend.load_dataset("mixed_test")
 
-        # JSONL should preserve all data types
-        assert loaded_data[0]["str_field"] == "text"
-        assert loaded_data[0]["int_field"] == 42  # Should be int
-        assert loaded_data[0]["float_field"] == 3.14  # Should be float
-        assert loaded_data[0]["bool_field"] is True  # Should be bool
-        assert loaded_data[0]["null_field"] is None  # Should be None
+        # All values become strings in CSV - these assertions should fail
+        assert loaded_data[0]["str_field"] == "text"  # This works
+        assert loaded_data[0]["int_field"] == 42  # This will fail - it's "42" not 42
+        assert (
+            loaded_data[0]["float_field"] == 3.14
+        )  # This will fail - it's "3.14" not 3.14
+        assert (
+            loaded_data[0]["bool_field"] is True
+        )  # This will fail - it's "True" not True
 
+    @pytest.mark.skip(reason="CSV backend doesn't support datetime objects")
     def test_datetime_objects(self, backend):
-        """Test datetime serialization - JSONL should handle this with ISO format."""
+        """Test datetime serialization - THIS SHOULD FAIL."""
         data = [
             {
                 "id": 1,
@@ -264,25 +275,16 @@ class TestDataTypeEdgeCases:
         backend.save_dataset("datetime_test", data)
         loaded_data = backend.load_dataset("datetime_test")
 
-        # JSONL should either preserve datetime objects or convert to ISO strings
-        # For now, let's expect ISO strings that can be parsed back
+        # Datetime objects become string representations - this should fail
         original_dt = data[0]["created_at"]
         loaded_dt = loaded_data[0]["created_at"]
 
-        # Should be either datetime object or ISO string
         assert isinstance(original_dt, datetime)
-        if isinstance(loaded_dt, str):
-            # If string, should be valid ISO format
-            parsed_dt = datetime.fromisoformat(loaded_dt.replace("Z", "+00:00"))
-            assert parsed_dt.year == 2024
-            assert parsed_dt.month == 1
-            assert parsed_dt.day == 15
-        else:
-            # If datetime object, should be exact match
-            assert loaded_dt == original_dt
+        assert isinstance(loaded_dt, datetime)  # This will fail - it's a string now!
 
+    @pytest.mark.skip(reason="CSV backend doesn't support complex nested structures")
     def test_complex_nested_structure(self, backend):
-        """Test deeply nested structures - JSONL should handle this perfectly."""
+        """Test deeply nested structures - THIS SHOULD FAIL BADLY."""
         data = [
             {
                 "config": {
@@ -299,12 +301,12 @@ class TestDataTypeEdgeCases:
         backend.save_dataset("complex_test", data)
         loaded_data = backend.load_dataset("complex_test")
 
-        # JSONL should preserve complex nested structures exactly
+        # This will fail - complex nested structure becomes string
         assert loaded_data[0]["config"]["database"]["host"] == "localhost"
-        assert loaded_data[0]["config"]["database"]["ports"] == [5432, 5433]
-        assert loaded_data[0]["config"]["database"]["credentials"]["user"] == "admin"
-        assert loaded_data[0]["config"]["database"]["credentials"]["encrypted"] is True
-        assert loaded_data[0]["config"]["features"] == ["auth", "logging"]
+
+        # Show the mangled data
+        print(f"Original: {data[0]['config']}")
+        print(f"Loaded: {loaded_data[0]['config']}")
 
 
 # 3. BaseModel Integration Tests
@@ -319,29 +321,29 @@ class TestBaseModelIntegration:
         # Load and validate with BaseModel
         loaded_data = backend.load_dataset("simple_model_test")
 
-        # JSONL should enable perfect BaseModel roundtrip
-        models = [SimpleTestModel(**item) for item in loaded_data]
-        assert len(models) == 3
-        assert models[0].name == "Alice"
-        assert models[0].age == 30
-        assert models[0].score == 85.5
-        assert models[0].is_active is True
+        # Try to create BaseModel instances - this will partially fail
+        try:
+            models = [SimpleTestModel(**item) for item in loaded_data]
+            print("BaseModel creation succeeded!")
+            print(f"First model: {models[0]}")
+        except Exception as e:
+            print(f"BaseModel creation failed: {e}")
+            print(
+                f"Loaded data types: {[(k, type(v)) for k, v in loaded_data[0].items()]}"
+            )
 
+    @pytest.mark.skip(reason="CSV backend doesn't support complex BaseModel validation")
     def test_complex_basemodel_roundtrip(self, backend, complex_data):
-        """Test BaseModel with complex data - JSONL should handle this."""
+        """Test BaseModel with complex data - THIS SHOULD FAIL."""
         # Save raw data
         backend.save_dataset("complex_model_test", complex_data, ComplexTestModel)
 
         # Load and try to validate
         loaded_data = backend.load_dataset("complex_model_test")
 
-        # JSONL should enable perfect BaseModel validation
-        models = [ComplexTestModel(**item) for item in loaded_data]
-        assert len(models) == 2
-        assert models[0].id == 1
-        assert models[0].metadata["score"] == 0.85
-        assert models[0].tags == ["evaluation", "metrics"]
-        assert models[0].config is not None and models[0].config["model"] == "gpt-4"
+        # This will fail because nested structures are corrupted
+        with pytest.raises(ValidationError):
+            [ComplexTestModel(**item) for item in loaded_data]
 
     def test_basemodel_type_coercion(self, backend):
         """Test BaseModel's ability to coerce string types."""
@@ -351,12 +353,12 @@ class TestBaseModelIntegration:
         backend.save_dataset("coercion_test", data)
         loaded_data = backend.load_dataset("coercion_test")
 
-        # JSONL + Pydantic should handle type coercion perfectly
+        # Pydantic should be able to handle some string-to-type conversions
+        # This might work for simple types
         model = SimpleTestModel(**loaded_data[0])
-        assert model.name == "Alice"
+        print(f"Type coercion successful: {model}")
         assert model.age == 30  # String "30" -> int 30
         assert model.score == 85.5  # String "85.5" -> float 85.5
-        # Note: "true" -> bool True coercion depends on implementation
 
 
 # 4. Error Handling & Edge Cases
@@ -385,30 +387,28 @@ class TestErrorHandling:
         backend.save_dataset("unicode_test", data)
         loaded_data = backend.load_dataset("unicode_test")
 
-        # Unicode should be preserved perfectly in JSONL
+        # Unicode should be preserved
         assert loaded_data[0]["name"] == "José María"
         assert loaded_data[0]["chinese"] == "你好世界"
         assert "🚀" in loaded_data[0]["description"]
 
-    def test_json_special_characters(self, backend):
-        """Test handling of JSON special characters."""
+    def test_csv_injection_protection(self, backend):
+        """Test protection against CSV injection attacks."""
+        # CSV injection attempts
         data = [
             {
-                "quotes": 'He said "Hello World"',
-                "backslashes": "C:\\Users\\test\\file.txt",
-                "newlines": "Line 1\nLine 2\nLine 3",
-                "tabs": "Column1\tColumn2\tColumn3",
+                "formula": "=SUM(A1:A10)",
+                "command": "@SUM(A1:A10)",
+                "plus_formula": "+SUM(A1:A10)",
+                "minus_formula": "-SUM(A1:A10)",
             }
         ]
 
-        backend.save_dataset("special_chars_test", data)
-        loaded_data = backend.load_dataset("special_chars_test")
+        backend.save_dataset("injection_test", data)
+        loaded_data = backend.load_dataset("injection_test")
 
-        # JSONL should handle JSON special characters properly
-        assert loaded_data[0]["quotes"] == 'He said "Hello World"'
-        assert loaded_data[0]["backslashes"] == "C:\\Users\\test\\file.txt"
-        assert loaded_data[0]["newlines"] == "Line 1\nLine 2\nLine 3"
-        assert loaded_data[0]["tabs"] == "Column1\tColumn2\tColumn3"
+        # Data should be preserved as-is (strings)
+        assert loaded_data[0]["formula"] == "=SUM(A1:A10)"
 
     def test_empty_and_null_values(self, backend):
         """Test handling of empty and null values."""
@@ -425,12 +425,10 @@ class TestErrorHandling:
         backend.save_dataset("empty_test", data)
         loaded_data = backend.load_dataset("empty_test")
 
-        # JSONL should handle null values properly
-        assert loaded_data[0]["empty_string"] == ""
-        assert loaded_data[0]["null_value"] is None
-        assert loaded_data[0]["whitespace"] == "   "
-        assert loaded_data[0]["zero"] == 0
-        assert loaded_data[0]["false"] is False
+        # Show how null values are handled
+        print(f"Original null: {data[0]['null_value']}")
+        print(f"Loaded null: {loaded_data[0]['null_value']}")
+        print(f"Loaded empty: '{loaded_data[0]['empty_string']}'")
 
     def test_large_text_fields(self, backend):
         """Test handling of large text fields."""
@@ -446,67 +444,24 @@ class TestErrorHandling:
         backend.save_dataset("large_text_test", data)
         loaded_data = backend.load_dataset("large_text_test")
 
-        # Large text should be preserved perfectly
+        # Large text should be preserved
         assert len(loaded_data[0]["large_field"]) == 10000
         assert loaded_data[0]["large_field"] == large_text
 
-    def test_malformed_jsonl_handling(self, backend, temp_dir):
-        """Test behavior with malformed JSONL files."""
-        # Create a malformed JSONL file manually
-        malformed_jsonl = Path(temp_dir) / "datasets" / "malformed.jsonl"
-        malformed_jsonl.parent.mkdir(parents=True, exist_ok=True)
+    def test_malformed_csv_handling(self, backend, temp_dir):
+        """Test behavior with malformed CSV files."""
+        # Create a malformed CSV file manually
+        malformed_csv = Path(temp_dir) / "datasets" / "malformed.csv"
+        malformed_csv.parent.mkdir(parents=True, exist_ok=True)
 
-        with open(malformed_jsonl, "w") as f:
-            f.write('{"valid": "json"}\n')
-            f.write('{"invalid": json}\n')  # Invalid JSON
-            f.write('{"another": "valid"}\n')
+        with open(malformed_csv, "w") as f:
+            f.write("header1,header2\n")
+            f.write("value1,value2,extra_value\n")  # Too many columns
+            f.write("value3\n")  # Too few columns
 
-        # Try to load malformed JSONL
+        # Try to load malformed CSV
         try:
             loaded_data = backend.load_dataset("malformed")
-            # Should either handle gracefully or raise appropriate error
-            print(f"Malformed JSONL loaded: {loaded_data}")
+            print(f"Malformed CSV loaded: {loaded_data}")
         except Exception as e:
-            print(f"Malformed JSONL failed to load: {e}")
-            # This is acceptable behavior
-
-
-# Helper functions for debugging
-def print_jsonl_content(jsonl_backend, data_type, name):
-    """Helper to print raw JSONL content for debugging."""
-    file_path = backend._get_file_path(data_type, name)
-    if file_path.exists():
-        print(f"\n=== JSONL Content for {name} ===")
-        with open(file_path, "r") as f:
-            print(f.read())
-        print("=== End JSONL Content ===\n")
-
-
-if __name__ == "__main__":
-    # Run some quick tests to see JSONL capabilities
-    import tempfile
-
-    with tempfile.TemporaryDirectory() as tmp_dir:
-        try:
-            backend: LocalJSONLBackend = LocalJSONLBackend(tmp_dir)
-
-            # Test nested data
-            test_nested_data: list[dict[str, t.Any]] = [
-                {"id": 1, "metadata": {"score": 0.85, "tags": ["test"]}}
-            ]
-            backend.save_dataset("debug_nested", test_nested_data)
-            loaded = backend.load_dataset("debug_nested")
-
-            print("=== Nested Data Test ===")
-            print(f"Original: {test_nested_data[0]['metadata']}")
-            print(f"Loaded: {loaded[0]['metadata']}")
-            print(
-                f"Types: {type(test_nested_data[0]['metadata'])} -> {type(loaded[0]['metadata'])}"
-            )
-
-            print_jsonl_content(backend, "datasets", "debug_nested")
-
-        except ImportError as e:
-            print(f"Expected ImportError: {e}")
-        except Exception as e:
-            print(f"Unexpected error: {e}")
+            print(f"Malformed CSV failed to load: {e}")
